Add sdk package
